+++ src/lib.rs
@@ -86,6 +86,7 @@ pub use friday::{
     FridayCitationRecord, FridayCodeTaskRecord, FridayCodeTaskStatus, FridayCompetitor,
     FridayConnectorAuthState, FridayConnectorKind, FridayConnectorRecord, FridayFeatureStatus,
     FridayGenerationSummary, FridayLocalCheckKind, FridayLocalCheckStatus,
+    FridayDeployGate, FridayBrowserVerificationReport, FridayBrowserVerificationTarget,
     FridayLocalExecutionCheck, FridayLocalExecutionReport, FridayMemoryRecord, FridayMemoryState,
     FridayMultimodalRuntimeRecord, FridayMultimodalSurface, FridayPermissionFinding,
     FridayPermissionScope, FridayPreviewRunner, FridayProductPlan, FridayProjectFile,
@@ -99,7 +100,8 @@ pub use friday::{
     FridayUiStateTone,
     FridayVoiceRuntimeRecord, FridayWorkspaceArea, FridayWorkspaceManifest,
     FridayWorkspaceSnapshot, FridayWorkspaceStore, FridayWorkspaceView, MetasearchExecutionMode,
-    MetasearchExecutionTarget, default_friday_local_execution_checks,
+    FridayVerificationStatus, MetasearchExecutionTarget,
+    default_friday_browser_verification_report, default_friday_local_execution_checks,
     default_friday_product_plan, default_friday_ui_integration_plan, friday_answer_search_plan,
     friday_research_search_plan,
 };